fix: Resolve race conditions in test container port allocation
